requirements.txt
openai==0.28
spacy
streamlit
nltk
gensim
wordcloud
python-dotenv